style(ecloud): remove unsed code & style (#219)
* style(api.go):remove unsed code

* style(ecloud)
