+++ pkg/cloud/ecloud/api.go
@@ -1,20 +1,3 @@
-/*
- * // Copyright 2022 CloudWeGo Authors
- * //
- * // Licensed under the Apache License, Version 2.0 (the "License");
- * // you may not use this file except in compliance with the License.
- * // You may obtain a copy of the License at
- * //
- * // http://www.apache.org/licenses/LICENSE-2.0
- * //
- * // Unless required by applicable law or agreed to in writing, software
- * // distributed under the License is distributed on an "AS IS" BASIS,
- * // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * // See the License for the specific language governing permissions and
- * // limitations under the License.
- * //
- */
-
 package ecloud
 
 import (
@@ -71,16 +54,6 @@ func (p *ECloud) ProviderType() string {
 }
 
 func (p *ECloud) DescribeImages(req cloud.DescribeImagesRequest) (cloud.DescribeImagesResponse, error) {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (p *ECloud) DescribeVpcs(req cloud.DescribeVpcsRequest) (cloud.DescribeVpcsResponse, error) {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (p *ECloud) DescribeSwitches(req cloud.DescribeSwitchesRequest) (cloud.DescribeSwitchesResponse, error) {
 	// TODO implement me
 	panic("implement me")
 }
