+++ pkg/cloud/ecloud/securitygroup.go
@@ -1,1 +1,23 @@
 package ecloud
+
+import "github.com/galaxy-future/BridgX/pkg/cloud"
+
+func (p *ECloud) CreateSecurityGroup(req cloud.CreateSecurityGroupRequest) (cloud.CreateSecurityGroupResponse, error) {
+	// TODO implement me
+	panic("implement me")
+}
+
+func (p *ECloud) AddIngressSecurityGroupRule(req cloud.AddSecurityGroupRuleRequest) error {
+	// TODO implement me
+	panic("implement me")
+}
+
+func (p *ECloud) AddEgressSecurityGroupRule(req cloud.AddSecurityGroupRuleRequest) error {
+	// TODO implement me
+	panic("implement me")
+}
+
+func (p *ECloud) DescribeSecurityGroups(req cloud.DescribeSecurityGroupsRequest) (cloud.DescribeSecurityGroupsResponse, error) {
+	// TODO implement me
+	panic("implement me")
+}
